agent-service/pyproject.toml
[project]
name = "nutrifit-agent-service"
version = "0.1.0"
description = "Agent service with FastAPI, Strands, and MCP integration"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "strands-agents>=1.0.0",
    "anthropic>=0.40.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "redis>=5.2.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["auth*", "agents*", "tools*"]
exclude = ["MCP*"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
